Realized toggle
A checkbox was added to the FlowGroup entries, allowing users to mark whether an expense has already been consolidated or not.
The calculation method for Estimated (and Budget in the FlowGroups table on the Dashboard) was modified: since we now have a boolean toggle to indicate when an expense has actually been realized, this indicator was also used for estimates. A sum of all expenses in the group (with realized = 0 or 1) was implemented. If this sum exceeds the budget, a warning is displayed on the FlowGroup screen indicating that the estimated expenses exceed the budget. This sum is now used as the basis for the Estimated value in the Balance section of the Dashboard.
Regarding income (salary), the realized button now behaves as follows: it helps calculate the Estimated value by summing entries with realized = 0 or 1 and displaying this total as Estimated. In the Income table on the Dashboard, entries that have not yet been realized are now shown in gray, while only values with realized = 1 remain in green.
Only values with realized = 1 are now summed to appear in the Balance as Realized.
The Income add button behaviour was changed. Now, once you click on it, it'll add the row directly to the Dashboard's Income Table.
Added a Realized button to the Dashboard's income table.